A bit of code cleaning up.
I got rid of unused variables in function signatures (in getLeft and
getRight, you only need the one side of the game).
I also changed your Show instance to use concatMap show instead of
foldl (++) "" (map show x), just for ease of readability.

I also used the @ "keyword" to refer to objects by a shorter name. This
came in really handy in the + declaration. It's much easier to
understand now.

I went ahead and added the rest of the Num methods just to get it to
compile; they're not defined so I just set the function body to
"undefined."

I got rid of (-). GHC understands that x - y is just x + negate y
(which is what you had), but it's unnecessary to explicitly state,
actually.

I also got rid of lots of unnecessary parenthesis. You are overusing
them a little bit, so I just cleaned that up a tad.